sampling completed but slow

+++ texasholdem/players.py
@@ -3,6 +3,6 @@ import numpy as np
 
 class TexasRange:
     def __init__(self, range_: str):
-        self.range_ = np.array([int(f) for f in range_.split()])
+        self.range_ = np.array([float(f) for f in range_.split()])
